remove unneeededd files
